# rest-stubs/go/api/swagger.yaml
---
swagger: "2.0"
info:
  version: "version not set"
  title: "todo.proto"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /v1/todo/records:
    get:
      tags:
      - "TodoService"
      operationId: "QueryTodo"
      parameters: []
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/TodoQueryRequest"
    post:
      tags:
      - "TodoService"
      operationId: "CreateTodo"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/TodoCreateRequest"
        x-exportParamName: "Body"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/TodoMutateResponse"
  /v1/todo/records/{id}:
    get:
      tags:
      - "TodoService"
      operationId: "GetTodo"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
        x-exportParamName: "Id"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/TodoGetResponse"
    put:
      tags:
      - "TodoService"
      operationId: "UpdateTodo"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
        x-exportParamName: "Id"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/TodoUpdateRequest"
        x-exportParamName: "Body"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/TodoMutateResponse"
    delete:
      tags:
      - "TodoService"
      operationId: "DeleteTodo"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
        x-exportParamName: "Id"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/TodoMutateResponse"
definitions:
  RecordStatus:
    type: "string"
    enum:
    - "STATUS_TODO"
    - "STATUS_INPROGRESS"
    - "STATUS_DONE"
    default: "STATUS_TODO"
  TodoCreateRequest:
    type: "object"
    properties:
      record:
        $ref: "#/definitions/TodoRecordInput"
    example:
      record:
        id: 0
        detail: "detail"
        title: "title"
        deadline: "2000-01-23T04:56:07.000+00:00"
  TodoGetResponse:
    type: "object"
    properties:
      record:
        $ref: "#/definitions/TodoRecord"
    example:
      record:
        created_at: "2000-01-23T04:56:07.000+00:00"
        id: 0
        detail: "detail"
        title: "title"
        deadline: "2000-01-23T04:56:07.000+00:00"
        status: {}
  TodoMutateResponse:
    type: "object"
    properties:
      record:
        $ref: "#/definitions/TodoRecord"
    example:
      record:
        created_at: "2000-01-23T04:56:07.000+00:00"
        id: 0
        detail: "detail"
        title: "title"
        deadline: "2000-01-23T04:56:07.000+00:00"
        status: {}
  TodoQueryRequest:
    type: "object"
    properties:
      start:
        type: "string"
        format: "date-time"
      end:
        type: "string"
        format: "date-time"
    example:
      start: "2000-01-23T04:56:07.000+00:00"
      end: "2000-01-23T04:56:07.000+00:00"
  TodoRecord:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      created_at:
        type: "string"
        format: "date-time"
      title:
        type: "string"
      detail:
        type: "string"
      deadline:
        type: "string"
        format: "date-time"
      status:
        $ref: "#/definitions/RecordStatus"
    example:
      created_at: "2000-01-23T04:56:07.000+00:00"
      id: 0
      detail: "detail"
      title: "title"
      deadline: "2000-01-23T04:56:07.000+00:00"
      status: {}
  TodoRecordInput:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      title:
        type: "string"
      detail:
        type: "string"
      deadline:
        type: "string"
        format: "date-time"
    example:
      id: 0
      detail: "detail"
      title: "title"
      deadline: "2000-01-23T04:56:07.000+00:00"
  TodoUpdateRequest:
    type: "object"
    properties:
      record:
        $ref: "#/definitions/TodoRecordInput"
      id:
        type: "integer"
        format: "int32"
    example:
      record:
        id: 0
        detail: "detail"
        title: "title"
        deadline: "2000-01-23T04:56:07.000+00:00"
      id: 0