# rest-stubs/go/todo_create_request.go
/*
 * todo.proto
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: version not set
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TodoCreateRequest struct {

	Record *TodoRecordInput `json:"record,omitempty"`
}