# rest-stubs/go/todo_get_response.go
/*
 * todo.proto
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: version not set
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TodoGetResponse struct {

	Record *TodoRecord `json:"record,omitempty"`
}