# rest-stubs/go/todo_query_request.go
/*
 * todo.proto
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: version not set
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"time"
)

type TodoQueryRequest struct {

	Start time.Time `json:"start,omitempty"`

	End time.Time `json:"end,omitempty"`
}