# rest-stubs/go/todo_update_request.go
/*
 * todo.proto
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: version not set
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TodoUpdateRequest struct {

	Record *TodoRecordInput `json:"record,omitempty"`

	Id int32 `json:"id,omitempty"`
}